+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.1.0
+
+- Improve template engine.
+- Improve authentication.
+- Add asset pipeline support.